split out operators
