+++ src/operator/index.ts
@@ -1,0 +1,8 @@
+// operator/index.ts
+
+export { default as forEach } from './for-each';
+export { default as map } from './map';
+export { default as filter } from './filter';
+export { default as complete } from './complete';
+export { default as error } from './error';
+export { default as split } from './split';
